Update fastapi requirement to account for 0.x (#179)
semver on 0.x.x version numbers treats minor updates equivalently to a
bump to 1.x.x; our requirement of fastapi at least 0.100.1 was not
compatible with the current newest version of 0.103.2. This PR changes
that bound to allow newest; we may need to revisit again for other
packages or when other 0.x dependencies update versions.