BUGFIX - mutable default arguments (#188)
Turns out Python doesn't create a new instance of a default mutable
argument each time you run `__init__` instead it will create the
resource once and then reuse it over and over. We found this bug while
working on the words_per_card dogfooding. It was causing all the decks
we read in from the file to contain every card from all decks. Oops!

This should fix it cleanly. I tested this in the average_words project.
Works fine. I've also added a test here to make sure it works.