#75 - Update utils.get_logger() (#248)
We want to use a Scooze logger throughout our code. This will be a
centralized way to get the logger and write to logs. We don't want to
use filters, formatters, or handlers for our logging because this is
library code and we want our end users to be able to determine how to
manage their own log output.